refactor: LoginController Annotation 기반으로 모두 수정 및 legacy Controller 제거
