+++ slipp/src/main/java/slipp/controller/LoginController.java
@@ -1,30 +1,46 @@
 package slipp.controller;
 
+import nextstep.mvc.tobe.ModelAndView;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.annotation.RequestMethod;
 import slipp.domain.User;
 import slipp.support.db.DataBase;
-import nextstep.mvc.asis.Controller;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-public class LoginController implements Controller {
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+@Controller
+public class LoginController {
+    @RequestMapping(value = "/users/loginForm", method = RequestMethod.GET)
+    public ModelAndView loginForm(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+        return ModelAndView.of("/user/login.jsp");
+    }
+
+    @RequestMapping(value = "/users/login", method = RequestMethod.POST)
+    public ModelAndView login(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         String userId = req.getParameter("userId");
         String password = req.getParameter("password");
         User user = DataBase.findUserById(userId);
         if (user == null) {
             req.setAttribute("loginFailed", true);
-            return "/user/login.jsp";
+            return ModelAndView.of("/user/login.jsp");
         }
         if (user.matchPassword(password)) {
             HttpSession session = req.getSession();
             session.setAttribute(UserSessionUtils.USER_SESSION_KEY, user);
-            return "redirect:/";
+            return ModelAndView.of("redirect:/");
         } else {
             req.setAttribute("loginFailed", true);
-            return "/user/login.jsp";
+            return ModelAndView.of("/user/login.jsp");
         }
+    }
+
+    @RequestMapping(value = "/users/logout", method = RequestMethod.GET)
+    public ModelAndView logout(HttpServletRequest req, HttpServletResponse resp) {
+        HttpSession session = req.getSession();
+        session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
+        return ModelAndView.of("redirect:/");
     }
 }
