--- slipp/src/main/java/slipp/controller/LoginUserController.java
@@ -1,35 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.tobe.ModelAndView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import slipp.domain.User;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-@Controller
-public class LoginUserController {
-
-    @RequestMapping(value = "/users/login", method = RequestMethod.POST)
-    public ModelAndView login(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        String userId = req.getParameter("userId");
-        String password = req.getParameter("password");
-        User user = DataBase.findUserById(userId);
-        if (user == null) {
-            req.setAttribute("loginFailed", true);
-            return ModelAndView.of("/user/login.jsp");
-        }
-        if (user.matchPassword(password)) {
-            HttpSession session = req.getSession();
-            session.setAttribute(UserSessionUtils.USER_SESSION_KEY, user);
-            return ModelAndView.of("redirect:/");
-        } else {
-            req.setAttribute("loginFailed", true);
-            return ModelAndView.of("/user/login.jsp");
-        }
-    }
-}
