--- slipp/src/main/java/slipp/controller/LogoutController.java
@@ -1,16 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.asis.Controller;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-public class LogoutController implements Controller {
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        HttpSession session = req.getSession();
-        session.removeAttribute(UserSessionUtils.USER_SESSION_KEY);
-        return "redirect:/";
-    }
-}
